+++ .gitignore
@@ -1,5 +1,8 @@
+# Ignore folders and files
 log/**
 .log*
 *.log
+# Folder
+log/*.log
 node_modules/
 .idea
