--- log/error.log
@@ -1,8 +1,0 @@
-[2020-10-01T15:06:32.879] [ERROR] debugError - Cheese is too ripe!
-[2020-10-01T15:06:32.879] [FATAL] debugError - Cheese was breeding ground for listeria.
-[2020-10-01T15:07:05.101] [ERROR] debugError - Cheese is too ripe!
-[2020-10-01T15:07:05.102] [FATAL] debugError - Cheese was breeding ground for listeria.
-[2020-10-01T15:07:13.934] [ERROR] debugError - Cheese is too ripe!
-[2020-10-01T15:07:13.934] [FATAL] debugError - Cheese was breeding ground for listeria.dsfsd
-
-sdfs
