+++ package.json
@@ -1,48 +1,71 @@
 {
-  "name": "api-backend",
-  "version": "1.0.0",
-  "description": "API REST ",
-  "author": "Marlon Zayro Arias Vargas",
-  "private": true,
-  "engines": {
-    "node": "14.x",
-    "npm": "6.x"
-  },
-  "scripts": {
-    "test": "mocha test/",
-    "test-markdown": "mocha test/ -R markdown > doc/spec.md --recursive",
-    "test-html": "mocha test/ -R doc > doc/spec.html --recursive",
-    "start": "node index.js",
-    "start-dev": "nodemon index.js"
-  },
-  "license": "ISC",
-  "dependencies": {
-    "bcrypt": "^3.0.6",
-    "body-parser": "^1.19.0",
-    "colors": "^1.3.0",
-    "compression": "^1.7.1",
-    "cors": "^2.8.4",
-    "dotenv": "^5.0.1",
-    "errorhandler": "^1.5.0",
-    "express": "^4.16.1",
-    "express-fileupload": "^0.3.0",
-    "jsonwebtoken": "^8.5.1",
-    "knex": "^0.21.6",
-    "log4js": "^6.3.0",
-    "method-override": "^2.1.3",
-    "moment": "^2.29.1",
-    "mongoose": "^5.10.8",
-    "mysql": "^2.18.1",
-    "pg": "^8.2.1",
-    "sha256": "^0.2.0",
-    "swagger-ui-express": "^4.0.7"
-  },
-  "devDependencies": {
-    "chai": "^4.1.2",
-    "chai-http": "^4.0.0",
-    "mocha": "^5.2.0",
-    "mochawesome": "^6.1.1",
-    "nodemon": "^1.19.4",
-    "supertest": "^3.0.0"
-  }
+    "name": "api-backend",
+    "version": "1.0.0",
+    "description": "API REST ",
+    "author": "Marlon Zayro Arias Vargas",
+    "private": true,
+    "engines": {
+        "node": "14.x",
+        "npm": "6.x"
+    },
+    "scripts": {
+        "test": "mocha test/",
+        "test-markdown": "mocha test/ -R markdown > doc/spec.md --recursive",
+        "test-html": "mocha test/ -R doc > doc/spec.html --recursive",
+        "start": "babel-node index.js  --exec ",
+        "watch": "babel src -d app --source-maps --watch",
+        "start-dev": "nodemon index.js --exec babel-node",
+        "build": "babel serve -s -d dist/",
+        "production": "npm run build && node dist/"
+    },
+    "license": "ISC",
+    "dependencies": {
+        "bcrypt": "^3.0.6",
+        "body-parser": "^1.19.0",
+        "colors": "^1.3.0",
+        "compression": "^1.7.1",
+        "cors": "^2.8.4",
+        "csurf": "^1.11.0",
+        "dotenv": "^5.0.1",
+        "errorhandler": "^1.5.0",
+        "express": "^4.16.1",
+        "express-fileupload": "^0.3.0",
+        "express-jsonschema": "^1.1.6",
+        "express-jwt": "^6.0.0",
+        "express-jwt-permissions": "^1.3.3",
+        "express-pdf": "^1.2.2",
+        "express-redis-cache": "^1.1.3",
+        "express-validator": "^6.10.0",
+        "helmet": "^4.5.0",
+        "jsonwebtoken": "^8.5.1",
+        "knex": "^0.21.6",
+        "method-override": "^2.1.3",
+        "moment": "^2.29.1",
+        "mongoose": "^5.10.8",
+        "morgan": "^1.10.0",
+        "multer": "^1.4.2",
+        "mysql": "^2.18.1",
+        "numeraljs": "^1.5.6",
+        "pg": "^8.2.1",
+        "pm2": "^4.5.6",
+        "sha256": "^0.2.0",
+        "swagger-ui-express": "^4.0.7"
+    },
+    "devDependencies": {
+        "@babel/cli": "^7.13.16",
+        "@babel/core": "^7.13.16",
+        "@babel/node": "^7.13.13",
+        "@babel/plugin-proposal-optional-chaining": "^7.13.12",
+        "@babel/plugin-proposal-pipeline-operator": "^7.12.13",
+        "@babel/preset-env": "^7.13.15",
+        "chai": "^4.1.2",
+        "chai-http": "^4.0.0",
+        "chalk": "^4.1.0",
+        "faker": "^5.5.3",
+        "longjohn": "^0.2.12",
+        "mocha": "^5.2.0",
+        "mochawesome": "^6.1.1",
+        "nodemon": "^1.19.4",
+        "supertest": "^3.0.0"
+    }
 }
