+++ src/.babelrc
@@ -1,0 +1,4 @@
+{
+    "presets": ["@babel/preset-env"]
+    //"ignore": ["node_modules"]
+}
