+++ src/db/index.js
@@ -1,5 +1,5 @@
 const knex = require('knex');
-const config = require('../config')
+const config = require('../config/config')
 
 const connection = knex(config.db);
