+++ src/index.js
@@ -1,7 +1,7 @@
 require('longjohn')
 
-const app = require('./src/app.js')
-import config from './src/config/config.js'
+const app = require('./app')
+const config = require('./config/config')
 
 app.listen(config.port, () => {
     console.log(`API REST corriendo en el puerto ${config.port}`)
