+++ src/model/general.js
