+++ src/test/index.js
@@ -1,6 +1,6 @@
 //apiTest.js
 const request = require('supertest');
-const app = require('../app');
+const app = require('../../app');
 const expect = require('chai').expect;
 const knex = require('../db')
 
@@ -8,7 +8,7 @@ const knex = require('../db')
 
 before(async () => {
     await console.log('antes de ...');
-     
+
 });
 
 after(async () => {
@@ -25,10 +25,10 @@ describe('GET /cliente', function () {
             .set('Accept', 'application/json')
             .expect('Content-Type', /json/)
             .expect(200)
-            .end(function(err, res) {
-              if (err) throw err;              
-              expect(res.status).to.equal(200);
-              done()
+            .end(function (err, res) {
+                if (err) throw err;
+                expect(res.status).to.equal(200);
+                done()
             });
     });
 });
@@ -36,40 +36,40 @@ describe('GET /cliente', function () {
 
 
 
-describe('POST /cliente',  () => {
-    it('respond with json containing to create support ticket', function (done)  {
-             request(app)
+describe('POST /cliente', () => {
+    it('respond with json containing to create support ticket', function (done) {
+        request(app)
             .post('/api/v1/client')
             .send({
                 "identificacion_cliente": "102030",
                 "descripcion": "arrego del tv"
-            })            
+            })
             .set('Accept', 'application/json')
             .expect('Content-Type', /json/)
             .expect(201)
-            .end(function(err, res) {
-                if (err) throw err;                            
+            .end(function (err, res) {
+                if (err) throw err;
                 done()
-              });
-            
+            });
+
     });
 });
 
 
-describe('PUT /cliente/1',  () => {
-    it('respond with json containing to create support ticket', function (done)  {
-             request(app)
+describe('PUT /cliente/1', () => {
+    it('respond with json containing to create support ticket', function (done) {
+        request(app)
             .put('/api/v1/cliente/1')
             .send({
                 "calificacion": 10
-            })            
+            })
             .set('Accept', 'application/json')
             .expect('Content-Type', /json/)
             .expect(200)
-            .end(function(err, res) {
-                if (err) throw err;                            
+            .end(function (err, res) {
+                if (err) throw err;
                 done()
-              });
-            
+            });
+
     });
 });
